requirements.txt
click==8.1.8
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
SQLAlchemy==2.0.37